Handle cases where the k8s namespace collides with env vars and make it possible to deploy on k8s and docker

+++ quobyte-docker-templates/main.sh
@@ -1,11 +1,10 @@
 #!/bin/bash
-# Set S3 Endpoint
-_S3=s3.quobyte.local
 
-MYNAME=$NODENAME
-STRIPPEDNAME=$(echo "$MYNAME" | tr -dc "a-z-")
-NODENUM=$(echo "$MYNAME" | tr -dc "0-9")
-
+if [ ! -z $CLOUD_INSTALL ]; then
+  MYNAME=$NODENAME
+  STRIPPEDNAME=$(echo "$MYNAME" | tr -dc "a-z-")
+  NODENUM=$(echo "$MYNAME" | tr -dc "0-9")
+fi
 
 function replaceOrAddParam () {
     local config_file=$1
@@ -21,51 +20,62 @@ function replaceOrAddParam () {
 }
 
 
-uname -a
+if [ ! -z $CLOUD_INSTALL ]; then
+  uname -a
 
-if [ "$MYNAME" == "quobyte-reg-0" ];then
-    if [ -e /var/lib/quobyte/devices/registry/QUOBYTE_DEV_SETUP ];then 
-        echo "registry exists"
-    else
-        /usr/bin/qbootstrap -y -d /var/lib/quobyte/devices/registry
-    fi
-else
-    if [ "$STRIPPEDNAME" == "quobyte-reg-" ];then
-        if [ -e /var/lib/quobyte/devices/registry/QUOBYTE_DEV_SETUP ];then 
-            echo "registry exists"
-        elif 
-            [[ $NODENUM < 5 ]]; then
-            /usr/bin/qmkdev -d -t REGISTRY /var/lib/quobyte/devices/registry
+  if [ "$MYNAME" == "quobyte-reg-0" ];then
+      if [ -e /var/lib/quobyte/devices/registry/QUOBYTE_DEV_SETUP ];then 
+          echo "registry exists"
+      else
+          /usr/bin/qbootstrap -y -d /var/lib/quobyte/devices/registry
+      fi
+  else
+      if [ "$STRIPPEDNAME" == "quobyte-reg-" ];then
+          if [ -e /var/lib/quobyte/devices/registry/QUOBYTE_DEV_SETUP ];then 
+              echo "registry exists"
+          elif 
+              [[ $NODENUM < 5 ]]; then
+              /usr/bin/qmkdev -d -t REGISTRY /var/lib/quobyte/devices/registry
+          fi
+      fi
+  fi
+
+  if [ "$STRIPPEDNAME" == "quobyte-meta-" ];then
+      for i in $(basename $(ls -d /var/lib/quobyte/devices/metadata*))
+      do
+        if [ -e /var/lib/quobyte/devices/${i}/QUOBYTE_DEV_SETUP ];then 
+          echo "metadata exists"
+        else
+          /usr/bin/qmkdev -d -t METADATA /var/lib/quobyte/devices/${i}
         fi
-    fi
+      done
+  fi
+
+  if [ "$STRIPPEDNAME" == "quobyte-data-" ];then
+      for i in $(basename $(ls -d /var/lib/quobyte/devices/data*))
+      do
+      if [ -e /var/lib/quobyte/devices/${i}/QUOBYTE_DEV_SETUP ];then 
+          echo "data exists"
+      else
+          /usr/bin/qmkdev -d -t DATA /var/lib/quobyte/devices/${i}
+      fi
+      done
+  fi
+
+  if [ "$STRIPPEDNAME" == "quobyte-web-" ];then
+      echo "constants.webconsole.setup_wizard.enable=false" > /etc/quobyte/webconsole.cfg
+  fi
+
+  if [ -z $QUOBYTE_WEBCONSOLE_PORT ]; then
+    QUOBYTE_WEBCONSOLE_PORT=8080
+  else
+   if [ $(echo $QUOBYTE_WEBCONSOLE_PORT | grep -c "^tcp") -eq 0 ]; then
+     QUOBYTE_WEBCONSOLE_PORT=$(echo $QUOBYTE_WEBCONSOLE_PORT | awk -F: '{print $3}')
+   fi 
+  fi
+
+  echo "registry=quobyte-reg-0.quobyte.${CLUSTER_NS}.svc.cluster.local,quobyte-reg-1.quobyte.${CLUSTER_NS}.svc.cluster.local,quobyte-reg-2.quobyte.${CLUSTER_NS}.svc.cluster.local,quobyte-reg-3.quobyte.${CLUSTER_NS}.svc.cluster.local" > /etc/quobyte/host.cfg
 fi
-
-if [ "$STRIPPEDNAME" == "quobyte-meta-" ];then
-    if [ -e /var/lib/quobyte/devices/metadata/QUOBYTE_DEV_SETUP ];then 
-        echo "metadata exists"
-    else
-        /usr/bin/qmkdev -d -t METADATA /var/lib/quobyte/devices/metadata
-    fi
-fi
-
-if [ "$STRIPPEDNAME" == "quobyte-data-" ];then
-    for i in $(basename $(ls -d /var/lib/quobyte/devices/data*))
-    do
-    if [ -e /var/lib/quobyte/devices/${i}/QUOBYTE_DEV_SETUP ];then 
-        echo "data exists"
-    else
-        /usr/bin/qmkdev -d -t DATA /var/lib/quobyte/devices/${i}
-    fi
-    done
-fi
-
-if [ "$STRIPPEDNAME" == "quobyte-web-" ];then
-    echo "constants.webconsole.setup_wizard.enable=false" > /etc/quobyte/webconsole.cfg
-fi
-
-QUOBYTE_WEBCONSOLE_PORT=8080
-
-echo "registry=quobyte-reg-0.quobyte.${CLUSTER_NS}.svc.cluster.local,quobyte-reg-1.quobyte.${CLUSTER_NS}.svc.cluster.local,quobyte-reg-2.quobyte.${CLUSTER_NS}.svc.cluster.local,quobyte-reg-3.quobyte.${CLUSTER_NS}.svc.cluster.local" > /etc/quobyte/host.cfg
 
 if [ -n "$QUOBYTE_RPC_PORT" ]; then echo rpc.port=$QUOBYTE_RPC_PORT > /etc/quobyte/$QUOBYTE_SERVICE.cfg; fi
 if [ -n "$QUOBYTE_HTTP_PORT" ]; then echo http.port=$QUOBYTE_HTTP_PORT >> /etc/quobyte/$QUOBYTE_SERVICE.cfg; fi
@@ -92,8 +102,10 @@ if [ -n "$QUOBYTE_MIN_MEM_API" ]; then replaceOrAddParam "/etc/default/quobyte" 
 if [ -n "$QUOBYTE_MIN_MEM_WEBCONSOLE" ]; then replaceOrAddParam "/etc/default/quobyte" "MIN_MEM_WEBCONSOLE" "$QUOBYTE_MIN_MEM_WEBCONSOLE"; fi
 if [ -n "$QUOBYTE_MIN_MEM_S3" ]; then replaceOrAddParam "/etc/default/quobyte" "MIN_MEM_S3" "$QUOBYTE_MIN_MEM_S3"; fi
 
-touch /etc/quobyte/$QUOBYTE_SERVICE.cfg
-sed -i -n -e '/^public_ip=/!p' -e "\$apublic_ip=$MY_POD_IP" /etc/quobyte/$QUOBYTE_SERVICE.cfg
+if [ ! -z $CLOUD_INSTALL ]; then
+  touch /etc/quobyte/$QUOBYTE_SERVICE.cfg
+  sed -i -n -e '/^public_ip=/!p' -e "\$apublic_ip=$MY_POD_IP" /etc/quobyte/$QUOBYTE_SERVICE.cfg
+fi
 
 SERVICE_UUID=$(grep "^ *uuid" "/etc/quobyte/$QUOBYTE_SERVICE.cfg" | awk -F = '{print $2}')
 if [[ -z $SERVICE_UUID ]]; then
@@ -105,37 +117,63 @@ if [[ -z $SERVICE_UUID ]]; then
 fi
 
 # The limits are passed by helm to the yaml file
-#export LIMIT_OPEN_FILES=1048576
-#export LIMIT_MAX_PROCESSES=16384
+if [ -z $CLOUD_INSTALL ]; then
+  export LIMIT_OPEN_FILES=1048576
+  export LIMIT_MAX_PROCESSES=16384
+fi
 
 ulimit -n $LIMIT_OPEN_FILES
 # Maximize the virtual memory limit to make sure that Java can set the MaxHeapSize (-Xmx) correctly.
 ulimit -v unlimited
 ulimit -u $LIMIT_MAX_PROCESSES
 
-# The service is defined for the container in the yaml file
-/usr/bin/quobyte-$QUOBYTE_SERVICE &
+if [ ! -z $CLOUD_INSTALL ]; then
+  if [ "$STRIPPEDNAME" != "quobyte-reg-" ];then
+    while true
+    do
+    j=0
+    for i in $(seq 0 3)
+    do
+      reg=$(dig +short quobyte-reg-${i}.quobyte.${CLUSTER_NS}.svc.cluster.local)
+      if [[ -z $reg ]]; then
+         echo "Waiting for quobyte-reg-${i}.quobyte.${CLUSTER_NS}.svc.cluster.local"
+         sleep 1
+      else
+        ((j++))
+      fi
+    done
 
-if [ "$MYNAME" == "quobyte-reg-0" ];then
-  /usr/bin/qmgmt -r user login admin quobyte
-  while [[ $(/usr/bin/qmgmt device list | grep -c "not registered") != 0 ]]
-  do 
-    echo "Waiting for devices to be registered"
-    sleep 1
-  done
+    if [[ $j -gt 3 ]]; then
+      break
+    fi
+    done
+  fi
+  # The service is defined for the container in the yaml file
+  /usr/bin/quobyte-$QUOBYTE_SERVICE &
 
-  /usr/bin/qmgmt volume config import RF1 /RF1.cfg
-  /usr/bin/qmgmt volume config import RF3 /RF3.cfg
-  /usr/bin/qmgmt volume config import EC42 /EC42.cfg
-  /usr/bin/qmgmt volume config import S3-objects /S3-objects.cfg
+  if [ "$MYNAME" == "quobyte-reg-0" ];then
+    /usr/bin/qmgmt -r user login admin quobyte
+    while [[ $(/usr/bin/qmgmt device list | grep -c "not registered") != 0 ]]
+    do 
+      echo "Waiting for devices to be registered"
+      sleep 1
+    done
 
-  /usr/bin/qmgmt volume create vol-RF1 root root RF1
-  /usr/bin/qmgmt volume create vol-RF3 root root RF3
-  /usr/bin/qmgmt volume create vol-EC42 root root EC42
-  /usr/bin/qmgmt volume create "S3 Objects" root root S3-objects
+    /usr/bin/qmgmt volume config import RF1 /RF1.cfg
+    /usr/bin/qmgmt volume config import RF3 /RF3.cfg
+    /usr/bin/qmgmt volume config import EC42 /EC42.cfg
+    /usr/bin/qmgmt volume config import S3-objects /S3-objects.cfg
 
-  sed -i "s/__S3__/${_S3}/g" /system.cfg
-  /usr/bin/qmgmt systemconfig import /system.cfg
-  rm -f /*.cfg
+    /usr/bin/qmgmt volume create vol-RF1 root root RF1
+    /usr/bin/qmgmt volume create vol-RF3 root root RF3
+    /usr/bin/qmgmt volume create vol-EC42 root root EC42
+    /usr/bin/qmgmt volume create "S3 Objects" root root S3-objects
+
+    sed -i "s/__S3__/${_S3}/g" /system.cfg
+    /usr/bin/qmgmt systemconfig import /system.cfg
+    rm -f /*.cfg
+  fi
+  wait
+else
+  /usr/bin/quobyte-$QUOBYTE_SERVICE
 fi
-wait
